fix: разделил на два файла,пересобрал responsives
