+++ npm-responsive/src/index.js
@@ -1,0 +1,4 @@
+import { MediaQuery } from "./MediaQuery";
+import { useMediaQuery } from "./useMediaQuery";
+
+export { MediaQuery, useMediaQuery };
